Commit - Survey Transaction Api and Batch

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/SurveyTransactionReportBranch.java
@@ -19,10 +19,10 @@ public class SurveyTransactionReportBranch
     private String surveyTransactionReportBranchId;
     
     @Column ( name = "month")
-    private Long month;
+    private int month;
     
     @Column ( name = "year")
-    private Long year;
+    private int year;
     
     @Column ( name = "user_name")
     private String userName;
@@ -119,22 +119,22 @@ public class SurveyTransactionReportBranch
         this.surveyTransactionReportBranchId = surveyTransactionReportBranchId;
     }
 
-    public Long getMonth()
+    public int getMonth()
     {
         return month;
     }
 
-    public void setMonth( Long month )
+    public void setMonth( int month )
     {
         this.month = month;
     }
 
-    public Long getYear()
+    public int getYear()
     {
         return year;
     }
 
-    public void setYear( Long year )
+    public void setYear( int year )
     {
         this.year = year;
     }
